- Fix alpha not drawing properly - Backport Ares `ParticleSystem` optimization - Fix `DetonateParticleSystem` attached to `Firer` - Fix `BehavesLike` for `ParticleSystemTypeClass` is different from `ParticleTypeClass` - Fix some defnition mistakes on YRPP - Make all `VectorHelper` function `forceinline` - Remove duplicate helpers - Remove unnessesary `EnterBioReactorSound` and `LeaveBioReactorSound` - Fix `DrawRadialIndicator` hook crashes - Fix Techno calculating threat for victim that doesnt exist - Cleanups

+++ YRpp/ParticleSystemClass.h
@@ -76,7 +76,7 @@ public:
 	CoordStruct TargetCoords;
 	ObjectClass* Owner;
 	AbstractClass* Target; // CellClass or TechnoClass
-	int          SpawnFrames; //from ParSysTypeClass
+	float          SpawnFrames; //from ParSysTypeClass
 	int          Lifetime; //from ParSysTypeClass
 	int          SparkSpawnFrames; //from ParSysTypeClass
 	int          SpotlightRadius; //defaults to 29
